refactor: Rename feed package to buildconfig
