+++ src/main/java/org/opentripplanner/standalone/config/BuildConfig.java
@@ -26,12 +26,12 @@ import org.opentripplanner.model.calendar.ServiceDateInterval;
 import org.opentripplanner.netex.config.NetexFeedParameters;
 import org.opentripplanner.routing.api.request.RouteRequest;
 import org.opentripplanner.routing.fares.FareServiceFactory;
-import org.opentripplanner.standalone.config.feed.DemConfig;
-import org.opentripplanner.standalone.config.feed.NetexConfig;
-import org.opentripplanner.standalone.config.feed.OsmConfig;
-import org.opentripplanner.standalone.config.feed.TransferRequestConfig;
-import org.opentripplanner.standalone.config.feed.TransitFeedConfig;
-import org.opentripplanner.standalone.config.feed.TransitFeeds;
+import org.opentripplanner.standalone.config.buildconfig.DemConfig;
+import org.opentripplanner.standalone.config.buildconfig.NetexConfig;
+import org.opentripplanner.standalone.config.buildconfig.OsmConfig;
+import org.opentripplanner.standalone.config.buildconfig.TransferRequestConfig;
+import org.opentripplanner.standalone.config.buildconfig.TransitFeedConfig;
+import org.opentripplanner.standalone.config.buildconfig.TransitFeeds;
 import org.opentripplanner.standalone.config.framework.json.NodeAdapter;
 import org.opentripplanner.standalone.config.sandbox.DataOverlayConfigMapper;
 import org.opentripplanner.util.lang.ObjectUtils;
