+++ src/main/java/org/opentripplanner/standalone/config/buildconfig/TransitFeeds.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.standalone.config.feed;
+package org.opentripplanner.standalone.config.buildconfig;
 
 import java.net.URI;
 import java.util.List;
